feat: Refactor Contact component and extract form handling into separate components
