+++ src/components/sections/Contact/contact-data.ts
@@ -1,0 +1,15 @@
+import {z} from "zod";
+
+export const formSchema = z.object({
+    name: z.string().min(2, {
+        message: "Name must be at least 2 characters.",
+    }),
+    email: z.email({
+        message: "Please enter a valid email address.",
+    }),
+    message: z.string().min(10, {
+        message: "Message must be at least 10 characters.",
+    }),
+});
+
+
